refactor tiling related functions
